+++ client/e2e/configurations.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from './fixtures/fixtures';
+import { test, expect } from './fixtures';
 import { deleteAllConfigurations } from './db';
 
 test.describe('Configurations screen', () => {
